Modularizing client some more
